Remove the save step from lock screen and quick launch shortcuts
Hear me out on this: none of the settings pages in Android require an
explicit "save" step to be performed by the user. The change is always
made immediately after making a selection. Likewise, other CyanogenMod
screens such as the QS tiles also take effect immediately.

The save step when modifying lock screen slider targets and quick launch
shortcuts is simply put completely unnecessary and distracting. It also
makes it far too easy to accidentally lose all the changes the user has
made.

This commit gets rid of the save buttons and saves the changes
immediately.

Change-Id: I50d3cbd1cf80eafb9cd6b5ee1fca217f6ae6ae07